move language definitions into uiCommon
