+++ uiCommon/ide/datalogPowered/commonTheme.css
