+++ uiCommon/ide/datalogPowered/languages/basicBlocks/example.txt
