+++ uiCommon/ide/datalogPowered/languages/fp/example.txt
