+++ uiCommon/ide/datalogPowered/languages/json/example.txt
