+++ uiCommon/ide/datalogPowered/languages/json/json.grammar
