+++ uiCommon/ide/datalogPowered/languages/sql/example.txt
