+++ uiCommon/ide/datalogPowered/languages/sql/sql.dd.txt
