+++ uiCommon/ide/datalogPowered/languages/treeSQL/example.txt
